chore(cardwired): move functions to helpers.rs
