+++ crates/cardwire-daemon/src/analyzer/mod.rs
@@ -1,5 +1,6 @@
 //! Cardwire analyzer, only in analysis mode rn
 mod dynamic_analysis;
+mod helpers;
 mod models;
 mod static_analysis;
